implement Authorization

+++ Market/Controllers/UserController.cs
@@ -1,10 +1,12 @@
 ﻿using Market.Models;
 using Market.Models.Repositories;
 using Market.ViewModel;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Market.Controllers
 {
+    [Authorize]
     public class UserController : Controller
     {
         private readonly IUserRepository _userRepository;
@@ -31,26 +33,30 @@ namespace Market.Controllers
 
         public ActionResult EditUser(int id)
         {
-             var user =_userRepository.GetOne(id);
+            var user = _userRepository.GetOne(id);
             return View(user);
         }
 
         [HttpPost]
         public IActionResult EditUser(int id, UserViewModel user)
         {
-             
+
             _userRepository.Update(id, user);
             return RedirectToAction(nameof(Index));
         }
         public IActionResult Delete(int id)
         {
-            _userRepository.Delete(id);
+            string isAdmin = HttpContext.User.Claims.FirstOrDefault(x => x.Type == ApplicationClaimTypes.IsAdmin)!.Value;
+            if (Convert.ToBoolean(isAdmin))
+            {
+                _userRepository.Delete(id);
+            }
             return RedirectToAction(nameof(Index));
         }
         public IActionResult Search(string term)
         {
-            var result=_userRepository.Search(term);
-            return View("Index",result);
+            var result = _userRepository.Search(term);
+            return View("Index", result);
         }
 
     }
